src/requirement.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
pymysql==1.4.6